requirements.txt
google~=3.0.0
requests~=2.28.1
google-api-python-client~=2.55.0
Pillow~=9.2.0